+++ src/DropDownNormal.php
@@ -7,7 +7,6 @@ namespace GCWorld\Menu;
 class DropDownNormal
 {
     public string $id;
-    public ?string $default = null;
     /** @var array<string,array{id:string,name:string,obj:MenuPanel}> */
     protected array $panels = [];
 
@@ -57,15 +56,6 @@ class DropDownNormal
             $out .= '</div>';
         }
         return $out;
-    }
-
-    /**
-     * @param string $id
-     * @return void
-     */
-    public function setDefault(string $id): void
-    {
-        $this->default = $id;
     }
 
     /**